% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_divs.R
\name{add_divs}
\alias{add_divs}
\title{Add Divisions function}
\usage{
add_divs(x)
}
\description{
This function allows you to add a divisions variable using a substring of business unit.
}
\examples{
add_divs(x)
}
\keyword{divs}
