% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_divs.R
\name{create_divs}
\alias{create_divs}
\title{Create divisions function}
\usage{
create_divs(dir_path = getwd(), num = FALSE)
}
\description{
Create a directory and sub-directory for each ALDI division.
}
\examples{
create_divs(dir_path = getwd(), num = FALSE)
}
\keyword{divs}
