% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_names.R
\name{div_names}
\alias{div_names}
\title{Replace div numbers with div names}
\usage{
div_names(x)
}
\description{
This function allows you to replace a div number variable with the 3 letter abbreviated name.
}
\examples{
div_names(x)
}
\keyword{divs}
