% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_divs.R
\name{split_divs_csv}
\alias{split_divs_csv}
\title{Split each div group to .csv file in pre-created subdirectories}
\usage{
split_divs_csv(x, dir_path, file_name)
}
\description{
Call using do(split_divs_csv(., "P:/Directory", "Report Name"))
}
\examples{
do(split_divs_csv(., "P:/Directory", "Report Name"))
}
\keyword{divs}
